reduce last_cumulative_funding_rate to i64

+++ programs/clearing_house/src/state/events.rs
@@ -68,7 +68,7 @@ pub struct FundingPaymentRecord {
     pub market_index: u16,
     pub funding_payment: i64,
     pub base_asset_amount: i64,
-    pub user_last_cumulative_funding: i128,
+    pub user_last_cumulative_funding: i64,
     pub amm_cumulative_funding_long: i128,
     pub amm_cumulative_funding_short: i128,
 }
